[Update] Facade - Renamed ColorPaletteFacade to ColorPalette - Updated composer aliases - Updated TestCase
