+++ tests/TestCase.php
@@ -2,7 +2,6 @@
 
 namespace Kiritokatklian\LaravelColorPalette\Tests;
 
-use Illuminate\Foundation\AliasLoader;
 use Kiritokatklian\LaravelColorPalette\ColorPaletteServiceProvider;
 use Orchestra\Testbench\TestCase as Orchestra;
 
@@ -11,18 +10,13 @@ class TestCase extends Orchestra
     public function setUp(): void
     {
         parent::setUp();
-
-        $loader = AliasLoader::getInstance();
-        $loader->alias('ColorPalette', 'LaravelColorPalette\\Facades\\ColorPalette');
     }
 
     protected function getPackageProviders($app): array
     {
-        $serviceProviders = [
+        return [
             ColorPaletteServiceProvider::class,
         ];
-
-        return $serviceProviders;
     }
 
     protected function getEnvironmentSetUp($app)
